requirements.txt
flask
tensorflow
keras
numpy
pillow
opencv-python